services/scrapper/tiktok/pyproject.toml
[project]
name = "tiktok-scraper"
version = "0.2.0"
description = "TikTok scraper with compression and async upload support"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.0.0",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "minio>=7.1.0",
    "motor>=3.3.0",
    "playwright==1.55.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "zstandard>=0.22.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.148.7",
    "pytest-asyncio>=1.3.0",
]
